chore(sdk): Refactor old clients to own file
To prevents loops in package imports with new clients loading them
